KneeRecordViewModelクラスを作成、uiStateを定義、依存関係ののバージョンを一部変更、weatherの型をIntに変更
