+++ app/build.gradle.kts
@@ -1,6 +1,7 @@
 plugins {
     id("com.android.application")
     id("org.jetbrains.kotlin.android")
+    alias(libs.plugins.ksp)
 }
 
 android {
@@ -50,7 +51,9 @@ android {
 val compose_version: String by rootProject.extra
 
 dependencies {
-
+    implementation (project(":core:data"))
+    implementation (project(":core:database"))
+    implementation (project(":core:model"))
     implementation ("androidx.core:core-ktx:1.12.0")
     implementation ("org.jetbrains.kotlin:kotlin-bom:1.8.0")
     implementation ("androidx.lifecycle:lifecycle-runtime-ktx:2.7.0")
@@ -75,4 +78,9 @@ dependencies {
     debugImplementation ("androidx.compose.ui:ui-test-manifest")
     implementation ("androidx.compose.ui:ui-text-google-fonts:1.6.5")
     implementation ("androidx.compose.material:material-icons-extended:1.6.5")
+
+    implementation(libs.hilt.android)
+    ksp(libs.hilt.compiler)
+    ksp(libs.hilt.android.complier)
+    implementation(libs.activity.compose)
 }
