+++ app/src/main/java/com/example/kneediary/navigation/KneeDiaryNavHost.kt
@@ -1,8 +1,6 @@
 package com.example.kneediary.navigation
 
-import android.content.Context
 import androidx.compose.runtime.Composable
-import androidx.navigation.NavController
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.NavHost
 import androidx.navigation.compose.composable
@@ -12,7 +10,7 @@ import com.example.kneediary.ui.screens.navigated_screen.home_screen.date_screen
 import com.example.kneediary.ui.screens.navigated_screen.home_screen.note_screen.NoteScreen
 import com.example.kneediary.ui.screens.navigated_screen.home_screen.weekly_screen.WeeklyScreen
 import com.example.kneediary.ui.screens.record_screen.RecordNoteScreen
-import com.example.kneediary.ui.screens.record_screen.RecordScreen
+import com.example.kneediary.ui.screens.record_screen.record_screen.RecordScreen
 import com.example.kneediary.ui.screens.start_screen.StartScreen
 
 @Composable
