+++ core/data/src/main/java/com/repository/KneeRecordRepository.kt
@@ -1,9 +1,15 @@
 package com.repository
 
 import com.github.yumayuma708.apps.model.KneeRecord
-import java.util.Date
 
 //createメソッドは、KneeRecordオブジェクトを返す。
 interface KneeRecordRepository {
-    suspend fun create(date: Long, time: Long, isRight: Boolean, painLevel: Int, weather: Int, note: String): KneeRecord
+    suspend fun create(
+        date: Long,
+        time: Long,
+        isRight: Boolean,
+        painLevel: Int,
+        weather: String,
+        note: String
+    ): KneeRecord
 }
