+++ core/data/src/main/java/com/repository/LocalKneeRecordRepository.kt
@@ -12,7 +12,7 @@ class LocalKneeRecordRepository @Inject constructor(
     //privateで、このクラス内でしか使えないパラメータkneeRecordDaoを定義している。
     private val kneeRecordDao: KneeRecordDao,
 ): KneeRecordRepository {
-    override suspend fun create(date: Long, time: Long, isRight: Boolean, painLevel: Int, weather: Int, note: String): KneeRecord {
+    override suspend fun create(date: Long, time: Long, isRight: Boolean, painLevel: Int, weather: String, note: String): KneeRecord {
         //currentTimeMillis()メソッドは、1970年1月1日00:00:00 GMTからの経過時間をミリ秒単位で返している。値はLong型
         val currentTime = System.currentTimeMillis()
         //kneeRecordEntityのインスタンスを作成
