+++ core/model/bin/main/com/github/yumayuma708/apps/model/KneeRecord.kt
